requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Celery and Redis
celery==5.4.0
redis==5.2.1
flower==2.0.1

# AWS Bedrock
boto3==1.35.93
botocore==1.35.93

# Document Processing
extract-msg==0.52.0
pdf2image==1.17.0
Pillow==11.1.0
python-magic==0.4.27
python-magic-bin==0.4.14  # Windows binary

# PDF and Excel Generation
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.0

# Authentication and Security
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.6.0

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
